hyperon

Module metta

Contains MeTTa specific types, constants and functions.

Modules

interpreter

MeTTa assembly language implementation. See minimal MeTTa documentation for details.

runner

MeTTa Runner Implementation

text

MeTTa parser implementation.

types

Contains Rust functions working on types. All MeTTa specific constants are exported as a part of metta module.

Constants

ARROW_SYMBOL

ATOM_TYPE_ATOM

ATOM_TYPE_EXPRESSION

ATOM_TYPE_GROUNDED

ATOM_TYPE_SYMBOL

ATOM_TYPE_TYPE

ATOM_TYPE_UNDEFINED

ATOM_TYPE_VARIABLE

BAD_ARG_TYPE_SYMBOL

BAD_TYPE_SYMBOL

CALL_NATIVE_SYMBOL

CHAIN_SYMBOL

COLLAPSE_BIND_SYMBOL

CONS_ATOM_SYMBOL

CONTEXT_SPACE_SYMBOL

DECONS_ATOM_SYMBOL

EMPTY_SYMBOL

EQUAL_SYMBOL

ERROR_SYMBOL

EVALC_SYMBOL

EVAL_SYMBOL

FUNCTION_SYMBOL

HAS_TYPE_SYMBOL

INCORRECT_NUMBER_OF_ARGUMENTS_SYMBOL

METTA_SYMBOL

NOT_REDUCIBLE_SYMBOL

NO_RETURN_SYMBOL

RETURN_SYMBOL

STACK_OVERFLOW_SYMBOL

SUB_TYPE_SYMBOL

SUPERPOSE_BIND_SYMBOL

UNIFY_SYMBOL

UNIT_ATOM

UNIT_TYPE

Functions

atom_error_message

Returns a message string from an error expression

atom_is_error

Tests whether or not an atom is an error expression

error_atom

Initializes an error expression atom