hyperon::metta

Module interpreter

MeTTa assembly language implementation. See minimal MeTTa documentation for details.

Structs

InterpreterState

This wrapper is to keep interpreter interface compatible with previous implementation and will be removed in future. State of the interpreter which passed between interpret_step calls.

Functions

interpret

Interpret passed atom and return a new plan, result or error. This function blocks until result is calculated. For step by step interpretation one should use interpret_init and interpret_step functions.

interpret_init

Initialize interpreter and returns the starting interpreter state. See crate::metta::interpreter for algorithm explanation.

interpret_step

Perform next step of the interpretation return the resulting interpreter state. See crate::metta::interpreter for algorithm explanation.