hyperon::metta::runner

Module pkg_mgmt

Structs

DepEntry

A single entry in a PkgInfo’s dependencies, specifying the properties of a module that will satisfy a dependency

DirCatalog

Implements ModuleCatalog to load MeTTa modules from a file-system directory trying a number of FsModuleFormat formats in succession

DirModuleFmt

An object to identify and load a MeTTa module implemented as a directory

LocalCatalog

A ManagedCatalog type to mediate and aggregate the contents of one or more “upstream” catalogs

ModuleDescriptor

A data structure that uniquely identifies an exact instance of a module

PkgInfo

The object responsible for locating and selecting dependency modules for each MettaMod

SingleFileModuleFmt

An object to identify and load a single-file module (naked .metta files)

Enums

UpdateMode

An interface to facilitate direct programatic management of a catalog, usually as a local mirror of one or more remote catalogs used by a user to insulate them from upstream changes Indicates the desired behavior for updating the locally-cached module

Traits

FsModuleFormat

Implemented on a type to test if a given file-system path points to a MeTTa module, and to construct possible paths within a parent directory for a module of a certain name

ManagedCatalog

ModuleCatalog

Implemented for types capable of locating MeTTa modules

Functions

mod_name_from_url

Extracts the module name from a .git URL