hyperon::metta::runner::stdlib

Module module

Structs

BindOp

ImportOp

IncludeOp

ModSpaceOp

mod-space! returns the space of a specified module, loading the module if it’s not loaded already

PrintModsOp

This operation prints the modules loaded from the top of the runner