hyperon::metta

Module text

MeTTa parser implementation.

Structs

CharReader

Proxy type to allow constructing SExprParser from different types automatically. CharReader implements From for different types, while SExprParser::new inputs Into<CharReader<R>> argument.

SExprParser

Provides a parser for MeTTa code written in S-Expression Syntax

SyntaxNode

Tokenizer

Enums

SyntaxNodeType

The meaning of a parsed syntactic element, generated from a substring in the input text

Traits

Parser

Implemented on a type that yields atoms to be interpreted as MeTTa code. Typically by parsing source text