hyperon::metta::runner

Module modules

Structs

MettaMod

Contains state associated with a loaded MeTTa module

ModId

A reference to a MettaMod that is loaded into a Metta runner

Enums

Resource

Resource for loading

ResourceKey

Identifies a resource to retrieve from a ModuleLoader

Constants

MOD_NAME_SEPARATOR

The separator between parent and child module names in a module name path

SELF_MOD_NAME

The name to refer to the current module in the module name hierarchy

TOP_MOD_NAME

The name of the top module in a runner

Traits

ModuleLoader

Implemented to supply a loader functions for MeTTa modules